Fixed ID to start with 1 instead of 0. Fixed enums without type declaration.
Makefile and tools/vspec.py let ID begin with 0. However, ID should start with 1.

Some signals and attributes with enums did not have the data type declared which
caused them to be treated as branches. Branches, however, are not included in the
ID file and the fidl and vsi files.

Signed-off-by: Rudolf J Streif <rstreif@jaguarlandrover.com>